Cleared up documentation and moved static methods to utils.py
