+++ Scripts/Errors.py
@@ -24,5 +24,6 @@ class Errors(commands.Cog):
             await ctx.send(f'Error not supported. Visit https://github.com/BBArikL/BDBot to report the issue. '
                            f'The error is: {error}')
 
+
 def setup(client):  # Initialize the cog
     client.add_cog(Errors(client))
